requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
dataclasses==0.7
decorator==4.4.2
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.4
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==12.8
pyzbar==0.1.8
regex==2020.7.14
requests==2.24.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
SQLAlchemy==1.3.18
tokenizers==0.8.1rc1
torch==1.6.0
tornado==6.0.4
tqdm==4.48.0
transformers==3.0.2
urllib3==1.25.10
Werkzeug==1.0.1
